perf: move ValidateProps to utils
