+++ components/_util/validateProps.ts
@@ -1,0 +1,9 @@
+import { tuple } from './type';
+
+const ValidateStatuses = tuple('success', 'warning', 'error', 'validating', '');
+export type ValidateStatus = typeof ValidateStatuses[number];
+
+export type ValidateProps = {
+  validateStatus?: ValidateStatus;
+  hasFeedback?: boolean;
+};
