+++ components/date-picker/generatePicker/generateSinglePicker.tsx
@@ -21,15 +21,12 @@ import {
   Components,
 } from '.';
 import { PickerComponentClass } from './interface';
-import { ValidateStatus } from '../../form/FormItem';
 import { getInputValidationClassName } from '../../input/utils';
 import iconMap from '../../_util/validationIcons';
+import { ValidateProps, ValidateStatus } from '../../_util/validateProps';
 
 export default function generatePicker<DateType>(generateConfig: GenerateConfig<DateType>) {
-  type DatePickerProps = PickerProps<DateType> & {
-    validateStatus?: ValidateStatus;
-    hasFeedback?: boolean;
-  };
+  type DatePickerProps = PickerProps<DateType> & ValidateProps;
 
   function getPicker<InnerPickerProps extends DatePickerProps>(
     picker?: PickerMode,
